initial refactoring into a package
